+++ hoomd/mpcd/SlitPoreGeometryFillerGPU.cuh
@@ -37,7 +37,7 @@ cudaError_t slit_pore_draw_particles(Scalar4 *d_pos,
                                      const unsigned int first_idx,
                                      const Scalar kT,
                                      const uint64_t timestep,
-                                     const unsigned int seed,
+                                     const uint16_t seed,
                                      const unsigned int block_size);
 
 } // end namespace gpu
